client.py
from binance.client import Client
import os
from dotenv import load_dotenv

load_dotenv()

def get_client():
    api_key = os.getenv("API_KEY")
    api_secret = os.getenv("API_SECRET")

    client = Client(api_key, api_secret, testnet=True)  # ✅ IMPORTANT

    client.FUTURES_URL = "https://demo-fapi.binance.com"

    return client



