requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==26.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.0
dateparser==1.4.0
Django==5.2.3
frozenlist==1.8.0
idna==3.10
instagrapi==2.1.3
multidict==6.7.1
pillow==11.1.0
propcache==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pynput==1.7.7
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python_binance==1.0.36
pytz==2026.1.post1
regex==2026.4.4
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==16.0
yarl==1.23.0
rich==14.0.0
